Fixed `accessibility()` test: return value now is double

+++ r-package/tests/testthat/test-accessibility.R
@@ -185,7 +185,7 @@ test_that("output is correct", {
   # expect each column to be of right class
 
   expect_true(typeof(result_df_input$id ) == "character")
-  expect_true(typeof(result_df_input$accessibility) == "integer")
+  expect_true(typeof(result_df_input$accessibility) == "double")
 
 
   #  * r5r options ----------------------------------------------------------
